refactor DisplayedColors
